+++ src/main.js
@@ -45,12 +45,12 @@ import {
   idDocumentsVerboseFilter,
   localizedSaleTypesFilter,
   feeTypesFilter,
-  assetAmountFilter,
 } from './components/App/filters/filters'
 import { globalizeOperationType } from './components/App/filters/globalizeOperationType'
 import { globalizeRequestStateI } from './components/App/filters/globalizeRequestStateI'
 import { globalizeStatsOpType } from './components/App/filters/globalizeStatsOpType'
 import { globalizeLimitType } from './components/App/filters/globalizeLimitType'
+import { formatMoney } from './components/App/filters/formatMoney'
 
 async function init () {
   i18n.onLanguageChanged(lang => {
@@ -84,7 +84,7 @@ async function init () {
   Vue.filter('globalizeOperationType', globalizeOperationType)
   Vue.filter('globalizeStatsOpType', globalizeStatsOpType)
   Vue.filter('globalizeLimitType', globalizeLimitType)
-  Vue.filter('assetAmountFilter', assetAmountFilter)
+  Vue.filter('formatMoney', formatMoney)
 
   /* Init Sentry */
   ErrorTracker.init(params)